[s.d] Template serie poster and name and test popup
Question. Should we display the serie name considering the serie poster
should contain the serie name in most cases and it's not like you
wouldn't already know what the serie is considering you're coming from
the episode details page. I thought maybe just adding a tooltip to the
poster would be enough.

Also with the tooltip, I used this other version called popup which
seems to be a sexier version of tooltip but the only reason I used it
hear was because tooltip doesnt like loading on the left side, the CSS
must be stuffed up and I can't get the tooltip to remain loaded so I can
inspect the CSS :/ so I just used this for now.